lnk_park_realese_new_album_fix
