+++ tasks/mpi/grudzin_k_all_reduce_student/src/ops_mpi.cpp
@@ -1,4 +1,4 @@
-#include "mpi/grudzin_k_all_reduce/include/ops_mpi.hpp"
+#include "mpi/grudzin_k_all_reduce_student/include/ops_mpi.hpp"
 
 #include <algorithm>
 #include <climits>
@@ -8,7 +8,7 @@
 #include <thread>
 #include <vector>
 
-bool grudzin_k_all_reduce_mpi::TestMPITaskSequential::pre_processing() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskSequential::pre_processing() {
   internal_order_test();
 
   rows = taskData->inputs_count[0];
@@ -27,13 +27,13 @@ bool grudzin_k_all_reduce_mpi::TestMPITaskSequential::pre_processing() {
   return true;
 }
 
-bool grudzin_k_all_reduce_mpi::TestMPITaskSequential::validation() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskSequential::validation() {
   internal_order_test();
   return ((taskData->inputs_count.size() >= 2 && taskData->inputs_count[0] != 0 && taskData->inputs_count[1] != 0) &&
           taskData->inputs_count[1] == taskData->outputs_count[0]);
 }
 
-bool grudzin_k_all_reduce_mpi::TestMPITaskSequential::run() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskSequential::run() {
   internal_order_test();
   for (int i = 0; i < rows; ++i) {
     for (int j = 0; j < colums; j++) {
@@ -48,7 +48,7 @@ bool grudzin_k_all_reduce_mpi::TestMPITaskSequential::run() {
   return true;
 }
 
-bool grudzin_k_all_reduce_mpi::TestMPITaskSequential::post_processing() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskSequential::post_processing() {
   internal_order_test();
   for (int i = 0; i < colums; i++) {
     reinterpret_cast<int*>(taskData->outputs[0])[i] = cnt_[i];
@@ -57,96 +57,12 @@ bool grudzin_k_all_reduce_mpi::TestMPITaskSequential::post_processing() {
 }
 
 //----------------------------------------------------------------------------
-bool grudzin_k_all_reduce_mpi::TestMPITaskBoostRealization::pre_processing() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskMyRealization::pre_processing() {
   internal_order_test();
   return true;
 }
 
-bool grudzin_k_all_reduce_mpi::TestMPITaskBoostRealization::validation() {
-  internal_order_test();
-  if (world.rank() == 0) {
-    return (((taskData->inputs_count.size() >= 2 && taskData->inputs_count[0] != 0 && taskData->inputs_count[1] != 0) &&
-             taskData->inputs_count[1] == taskData->outputs_count[0]));
-  }
-  return true;
-}
-
-bool grudzin_k_all_reduce_mpi::TestMPITaskBoostRealization::run() {
-  internal_order_test();
-  if (world.rank() == 0) {
-    rows = taskData->inputs_count[0];
-    colums = taskData->inputs_count[1];
-  }
-  boost::mpi::broadcast(world, rows, 0);
-  boost::mpi::broadcast(world, colums, 0);
-
-  int size = rows * colums;
-  int rest = size % world.size();
-  int delta = size / world.size();
-
-  if (rest != 0) {
-    delta++;
-  }
-
-  if (world.rank() == 0) {
-    input_.assign(delta * world.size(), INT_MAX);
-    std::copy(reinterpret_cast<int*>(taskData->inputs[0]), reinterpret_cast<int*>(taskData->inputs[0]) + size,
-              input_.begin());
-  }
-
-  std::vector<int> locmin(colums, INT_MAX);
-  local_input_.resize(delta);
-
-  int start = delta;
-  if (world.rank() == 0) {
-    for (int proc = 1; proc < world.size(); proc++) {
-      world.send(proc, 0, input_.data() + start, delta);
-      start += delta;
-    }
-  }
-
-  start = delta * world.rank();
-  if (world.rank() == 0) {
-    std::copy(input_.begin(), input_.begin() + delta, local_input_.begin());
-  } else {
-    world.recv(0, 0, local_input_.data(), delta);
-  }
-
-  for (size_t i = 0; i < local_input_.size(); ++i) {
-    if (locmin[(start + i) % colums] > local_input_[i]) {
-      locmin[(start + i) % colums] = local_input_[i];
-    }
-  }
-  res_.resize(colums, INT_MAX);
-  boost::mpi::all_reduce(world, locmin.data(), colums, res_.data(), boost::mpi::minimum<int>());
-  std::vector<int> local_cnt_(colums, 0);
-  for (size_t i = 0; i < local_input_.size(); ++i) {
-    if (res_[(start + i) % colums] == local_input_[i]) {
-      local_cnt_[(start + i) % colums]++;
-    }
-  }
-  cnt_.resize(colums, 0);
-  boost::mpi::reduce(world, local_cnt_.data(), colums, cnt_.data(), std::plus<>(), 0);
-  return true;
-}
-
-bool grudzin_k_all_reduce_mpi::TestMPITaskBoostRealization::post_processing() {
-  internal_order_test();
-  if (world.rank() == 0) {
-    for (int i = 0; i < colums; i++) {
-      reinterpret_cast<int*>(taskData->outputs[0])[i] = cnt_[i];
-    }
-  }
-  return true;
-}
-
-//------------------------------------------------------------------------
-bool grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::pre_processing() {
-  internal_order_test();
-  return true;
-}
-
-bool grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::validation() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskMyRealization::validation() {
   internal_order_test();
   if (world.rank() == 0) {
     return (((taskData->inputs_count.size() >= 2 && taskData->inputs_count[0] != 0 && taskData->inputs_count[1] != 0) &&
@@ -156,8 +72,9 @@ bool grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::validation() {
 }
 
 template <typename T>
-void grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::my_all_reduce(const boost::mpi::communicator& world,
-                                                                       const T* in_values, int n, T* out_values) {
+void grudzin_k_all_reduce_student_mpi::TestMPITaskMyRealization::my_all_reduce(const boost::mpi::communicator& world,
+                                                                               const T* in_values, int n,
+                                                                               T* out_values) {
   int root = world.rank();
   T* cv1 = nullptr;
   T* cv2 = nullptr;
@@ -200,7 +117,7 @@ void grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::my_all_reduce(const boo
   }
 }
 
-bool grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::run() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskMyRealization::run() {
   internal_order_test();
   if (world.rank() == 0) {
     rows = taskData->inputs_count[0];
@@ -259,7 +176,7 @@ bool grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::run() {
   return true;
 }
 
-bool grudzin_k_all_reduce_mpi::TestMPITaskMyRealization::post_processing() {
+bool grudzin_k_all_reduce_student_mpi::TestMPITaskMyRealization::post_processing() {
   internal_order_test();
   if (world.rank() == 0) {
     for (int i = 0; i < colums; i++) {
